Lab2/requirements.txt
python-docx==0.8.11
openpyxl==3.1.2
pytest == 8.0.0